Fix problem with daylight saving time
Fix a problem with daliy scheduler in a enviroment with a timezone that has daiylight saving time. 
Problem description: 
I used a cron pattern like this [TimerTrigger("0 30 23 * * *")] /* 23:00 every day */
On a Azure Web App Service that has the Timezone set to Europe Standard Time. 
The run at 24. March starts as expected  at 24. March 23:00.
The problem occurs on the next run it starts on 26. March at 03:00 instate 25. March at 23:00.