Initial work on updating docs for SDK 3.0.0
The first example in the README isn't up to date with the 3.0 release
of the WebJobs SDK - this commit changes it to reflect the more modern
usage of `HostBuilder`. It also *begins* the work of changing from the
older `Use` calls to `Add` calls for using SDK extensions.

The docs around ErrorTrigger, Core Extensions and Twilio SMS extensions
haven't been updated yet as they don't have any corresponding code in
3.0.